+++ src/test/components/ng-http-loader.component.spec.ts
@@ -14,7 +14,7 @@ import { async, ComponentFixture, fakeAsync, inject, TestBed, tick } from '@angu
 import { By } from '@angular/platform-browser';
 import { BehaviorSubject, forkJoin, Observable } from 'rxjs';
 import { NgHttpLoaderComponent } from '../../lib/components/ng-http-loader.component';
-import { PendingInterceptorServiceProvider } from '../../lib/services/pending-interceptor.service';
+import { PendingRequestsInterceptorProvider } from '../../lib/services/pending-requests-interceptor.service';
 import { SpinnerVisibilityService } from '../../lib/services/spinner-visibility.service';
 import { Spinkit, SPINKIT_COMPONENTS } from '../../lib/spinkits';
 
@@ -32,7 +32,7 @@ describe('NgHttpLoaderComponent', () => {
         TestBed.configureTestingModule({
             declarations: [NgHttpLoaderComponent, HostComponent, ...SPINKIT_COMPONENTS],
             imports: [HttpClientTestingModule],
-            providers: [PendingInterceptorServiceProvider]
+            providers: [PendingRequestsInterceptorProvider]
         })
             .compileComponents();
     }));
@@ -47,7 +47,7 @@ describe('NgHttpLoaderComponent', () => {
     });
 
     it('should create the ng-http-loader component with default values', () => {
-        spyOnProperty(component, 'isSpinnerVisible$')
+        spyOnProperty(component, 'isVisible$')
             .and.returnValue(new BehaviorSubject(true).asObservable());
         fixture.detectChanges();
 
@@ -73,7 +73,7 @@ describe('NgHttpLoaderComponent', () => {
     });
 
     it('should not set the colored class if background-color is defined', () => {
-        spyOnProperty(component, 'isSpinnerVisible$')
+        spyOnProperty(component, 'isVisible$')
             .and.returnValue(new BehaviorSubject(true).asObservable());
         component.backgroundColor = '#ff0000';
         fixture.detectChanges();
@@ -95,7 +95,7 @@ describe('NgHttpLoaderComponent', () => {
     });
 
     it('should be able to specify another known spinner', () => {
-        spyOnProperty(component, 'isSpinnerVisible$')
+        spyOnProperty(component, 'isVisible$')
             .and.returnValue(new BehaviorSubject(true).asObservable());
         component.spinner = Spinkit.skRotatingPlane;
         fixture.detectChanges();
@@ -109,7 +109,7 @@ describe('NgHttpLoaderComponent', () => {
     });
 
     it('should allow us to specify a custom background-color', () => {
-        spyOnProperty(component, 'isSpinnerVisible$')
+        spyOnProperty(component, 'isVisible$')
             .and.returnValue(new BehaviorSubject(true).asObservable());
         component.backgroundColor = '#ff0000';
         fixture.detectChanges();
@@ -135,15 +135,15 @@ describe('NgHttpLoaderComponent', () => {
             const secondRequest = httpMock.expectOne('/fake2');
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             firstRequest.flush({});
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             secondRequest.flush({});
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -152,11 +152,11 @@ describe('NgHttpLoaderComponent', () => {
             http.get('/fake').subscribe();
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -167,7 +167,7 @@ describe('NgHttpLoaderComponent', () => {
 
             http.get('/fake').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
         }
     )));
@@ -179,7 +179,7 @@ describe('NgHttpLoaderComponent', () => {
 
             http.get('/fake').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
         }
     )));
@@ -196,7 +196,7 @@ describe('NgHttpLoaderComponent', () => {
             }).subscribe();
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
         }
     )));
@@ -215,7 +215,7 @@ describe('NgHttpLoaderComponent', () => {
                 }
             ).subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/api/service?foo=bar').flush({});
         }
     )));
@@ -227,16 +227,16 @@ describe('NgHttpLoaderComponent', () => {
 
             http.get('/12345').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/12345').flush({});
 
             http.get('/fake').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -247,16 +247,16 @@ describe('NgHttpLoaderComponent', () => {
 
             http.post('/12345', null).subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/12345').flush({});
 
             http.get('/fake').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -271,16 +271,16 @@ describe('NgHttpLoaderComponent', () => {
                 }
             }).subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.expectOne('/12345').flush({});
 
             http.get('/fake').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
 
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -307,11 +307,11 @@ describe('NgHttpLoaderComponent', () => {
             const newComponent = newFixture.componentInstance;
 
             tick();
-            newComponent.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            newComponent.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             httpMock.expectOne('/fake').flush({});
 
             tick();
-            newComponent.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            newComponent.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             httpMock.verify();
         }
     )));
@@ -323,24 +323,24 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP request is pending for 1,999 seconds now
             tick(999);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP request is pending for 2 seconds now - the spinner will be visible
             tick(1);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is pending for 5 seconds now - the spinner is still visible
             tick(3000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is finally over, the spinner is hidden
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -351,12 +351,12 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP request is over, the spinner shouldn't be shown after debounceDelay terminated
             httpMock.expectOne('/fake').flush({});
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -367,7 +367,7 @@ describe('NgHttpLoaderComponent', () => {
 
             // the first HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the first HTTP request is over
             httpMock.expectOne('/fake').flush({});
@@ -377,16 +377,16 @@ describe('NgHttpLoaderComponent', () => {
 
             // the second HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the second HTTP request is over
             httpMock.expectOne('/fake2').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the spinner shouldn't be shown after debounceDelay terminated
             tick(2000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -398,23 +398,23 @@ describe('NgHttpLoaderComponent', () => {
 
             // both HTTP requests are pending for 1s now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the first HTTP request is over
             httpMock.expectOne('/fake').flush({});
 
             // the second HTTP request is pending for 2s now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the second HTTP request is over
             httpMock.expectOne('/fake2').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the spinner shouldn't be shown after debounceDelay terminated
             tick(3000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -433,53 +433,53 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP requests are pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP requests are pending for 1,999 seconds now
             tick(999);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP requests are pending for 2 seconds now - the spinner will be visible
             tick(1);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP requests are pending for 5 seconds now - the spinner is still visible
             tick(3000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the first HTTP request is finally over, the spinner is still visible
             firstRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second request is pending for 8 seconds now - the spinner is still visible
             tick(3000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is finally over, the spinner is hidden
             secondRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
     it('should be possible to manually show/hide the spinner', inject(
         [SpinnerVisibilityService], (spinner: SpinnerVisibilityService) => {
             spinner.show();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             spinner.hide();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     ));
 
     it('should be possible to manually show/hide the spinner in a Promise context', inject(
         [SpinnerVisibilityService], (spinner: SpinnerVisibilityService) => {
             spinner.show();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             Promise.resolve('resolved').then(() => {
                 spinner.hide();
-                component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+                component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
             }).catch(() => expect(true).toBeFalsy());
         }
     ));
@@ -489,28 +489,28 @@ describe('NgHttpLoaderComponent', () => {
         (spinner: SpinnerVisibilityService, http: HttpClient, httpMock: HttpTestingController) => {
             // we manually show the spinner
             spinner.show();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             // then an HTTP request is performed
             http.get('/fake').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request ends, but we want the spinner to be still visible
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             spinner.hide();
             // this time the spinner is not visible anymore
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
-            // the bypassPendingInterceptorService should be reset for next HTTP requests
+            // _forceByPass should be reset for next HTTP requests
             http.get('/fake2').subscribe();
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
             httpMock.expectOne('/fake2').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -521,32 +521,32 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is pending for 2 seconds now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is finally over, the spinner is still visible
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is over but the spinner is still visible after 3 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the spinner is still visible after 4 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the spinner is still visible after 4,999 seconds
             tick(999);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the spinner is not visible anymore after 5 seconds
             tick(1);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -557,24 +557,24 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is pending for 2 seconds now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is finally over, the spinner is still visible
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // 4 seconds after the HTTP request is over, the spinner is still visible
             tick(4000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the spinner is not visible anymore after 5 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -593,33 +593,33 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP requests are pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP requests are pending for 2 seconds now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the first HTTP request is finally over, the spinner is still visible
             firstRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is still pending after 3 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is still pending after 4 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is finally over too, the spinner is still visible
             secondRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // After 5 seconds, the spinner is hidden
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -638,37 +638,37 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP requests are pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP requests are pending for 2 seconds now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the first HTTP request is finally over, the spinner is still visible
             firstRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is still pending after 3 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is still pending after 4 seconds
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the second HTTP request is finally over too, the spinner is still visible
             secondRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // After 4 seconds, the spinner is still visible
             tick(4000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // After 5 seconds, the spinner is hidden
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -680,30 +680,30 @@ describe('NgHttpLoaderComponent', () => {
             const firstRequest = httpMock.expectOne('/fake');
 
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the first HTTP request is finally over, the spinner is still visible for at least 1 second
             firstRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // But 200 ms after the first HTTP request has finished, a second HTTP request is launched
             tick(200);
             http.get('/fake2').subscribe();
             const secondRequest = httpMock.expectOne('/fake2');
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // After 900ms, the spinner should
             // still be visible because the second HTTP request is still pending
             tick(900);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // 500 ms later, the second http request ends. The spinner should be hidden
             // Total time spent visible (1000+200+1400==2600 > minDuration)
             tick(500);
             secondRequest.flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -719,12 +719,12 @@ describe('NgHttpLoaderComponent', () => {
             const firstRequest = httpMock.expectOne('/fake');
 
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the first HTTP request is finally over, the spinner is still visible for at least 10ms
             firstRequest.flush({});
             tick(5);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // But 5 ms after the first HTTP request has finished, a second HTTP request has been launched
             runQuery$('/fake2').subscribe();
@@ -733,11 +733,11 @@ describe('NgHttpLoaderComponent', () => {
             // After 700ms, the second http request ends. The spinner is still visible
             tick(700);
             secondRequest.flush({});
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // 10ms later, the spinner should be  hidden (extraDuration)
             tick(10);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -748,16 +748,16 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 1 second now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is pending for 2 seconds now
             tick(1000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is finally over after 2 seconds, the spinner is hidden
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -765,10 +765,10 @@ describe('NgHttpLoaderComponent', () => {
         [SpinnerVisibilityService], (spinner: SpinnerVisibilityService) => {
             component.minDuration = 10000;
             spinner.show();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             spinner.hide();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     ));
 
@@ -776,10 +776,10 @@ describe('NgHttpLoaderComponent', () => {
         [SpinnerVisibilityService], (spinner: SpinnerVisibilityService) => {
             component.extraDuration = 10000;
             spinner.show();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             spinner.hide();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     ));
 
@@ -793,28 +793,28 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 0,5 second now - spinner not visible because debounce
             tick(500);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP request is pending for 1 second now - spinner visible
             tick(500);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is finally over, the spinner is still visible
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // after 3 seconds, the spinner is still visible
             tick(2000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // after 5,999 seconds, the spinner is still visible
             tick(2999);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // after 6 seconds (1s for debounce + 5s extra. duration), the spinner is hidden
             tick(1);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 
@@ -828,28 +828,28 @@ describe('NgHttpLoaderComponent', () => {
 
             // the HTTP request is pending for 0,5 second now - spinner not visible because debounce
             tick(500);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
 
             // the HTTP request is pending for 1 second now - spinner visible
             tick(500);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // the HTTP request is finally over, the spinner is still visible
             httpMock.expectOne('/fake').flush({});
             tick();
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // after 3 seconds, the spinner is still visible
             tick(2000);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // after 5,999 seconds, the spinner is still visible
             tick(2999);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeTruthy()).unsubscribe();
 
             // after 6 seconds (1s for debounce + 5s min. duration), the spinner is hidden
             tick(1);
-            component.isSpinnerVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
+            component.isVisible$.subscribe(v => expect(v).toBeFalsy()).unsubscribe();
         }
     )));
 });