+++ src/test/ng-http-loader.module.spec.ts
@@ -8,7 +8,7 @@
  */
 
 import { NgHttpLoaderModule } from '../lib/ng-http-loader.module';
-import { PendingInterceptorService } from '../lib/services/pending-interceptor.service';
+import { PendingRequestsInterceptor } from '../lib/services/pending-requests-interceptor.service';
 
 describe('NgHttpLoaderModule', () => {
     let ngHttpLoaderModule: NgHttpLoaderModule;
@@ -23,6 +23,6 @@ describe('NgHttpLoaderModule', () => {
 
     it('should create an instance with providers via forRoot()', () => {
         const ngHttpLoaderModuleWithProviders = NgHttpLoaderModule.forRoot();
-        expect(ngHttpLoaderModuleWithProviders.providers[0][0]['useExisting'].name).toEqual(PendingInterceptorService.name);
+        expect(ngHttpLoaderModuleWithProviders.providers[0][0]['useExisting'].name).toEqual(PendingRequestsInterceptor.name);
     });
 });